rename of DynaGrid to ReactGrid, cleanup and commenting
