+++ src/index.tsx
@@ -1,14 +1,17 @@
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
 import DynaGridDemo from "./demo/Views/DynaGridDemo"
+// TODO how to remove this?
 import 'babel-polyfill';
 
 // DO NOT MOVE
 // this index.tsx is required by react-scripts-ts
+
+
 ReactDOM.render(
   <DynaGridDemo />,
   document.getElementById('root') as HTMLElement
 );
 
 export default DynaGridDemo;
-export * from "./lib/Components/DynaGrid";
+export * from "./lib";
