+++ src/lib/CellTemplates/DateCellTemplate.tsx
@@ -2,11 +2,13 @@ import * as React from 'react';
 import { keyCodes } from '../Common/Constants';
 import { CellRenderProps, CellTemplate } from '../Common';
 
-export class TextCellTemplate implements CellTemplate<string> {
+export class DateCellTemplate implements CellTemplate<string> {
     readonly hasEditMode = true;
 
     validate(data: any): string {
-        return (typeof (data) === 'string') ? data : '';
+        const date_regex = /^\d{4}\-\d{2}\-\d{2}$/;
+        data = data ? data.toString().replace(/\s+/g, '') : data;
+        return (date_regex.test(data)) ? data : '';
     }
 
     textToCellData(text: string): string {
@@ -21,31 +23,31 @@ export class TextCellTemplate implements CellTemplate<string> {
         return { editable: true, cellData }
     }
 
-    customStyle: React.CSSProperties = { background: '#fff' };
+    customStyle: React.CSSProperties = {};
 
     renderContent: (props: CellRenderProps<string>) => React.ReactNode = (props) => {
         if (!props.isInEditMode)
             return props.cellData;
         const preserveValueKeyCodes = [0, keyCodes.ENTER];
         return <input
+            type='date'
             style={{
-                position: 'inherit',
                 width: '100%',
                 height: '100%',
                 padding: 0,
                 border: 0,
                 background: 'transparent',
                 fontSize: 14,
-                outline: 'none',
+                outline: 'none'
             }}
             ref={input => {
                 if (input) {
                     input.focus();
-                    input.setSelectionRange(input.value.length, input.value.length);
+                    // input.setSelectionRange(input.value.length, input.value.length);
                 }
             }}
             defaultValue={preserveValueKeyCodes.includes(props.lastKeyCode) ? props.cellData : ''}
-            onChange={e => props.onCellDataChanged ? props.onCellDataChanged(e.currentTarget.value) : null}
+            onChange={e => props.onCellDataChanged(e.currentTarget.value)}
             onCopy={e => e.stopPropagation()}
             onCut={e => e.stopPropagation()}
             onPaste={e => e.stopPropagation()}
