+++ src/lib/CellTemplates/TextCellTemplate.tsx
@@ -2,13 +2,11 @@ import * as React from 'react';
 import { keyCodes } from '../Common/Constants';
 import { CellRenderProps, CellTemplate } from '../Common';
 
-export class TimeCellTemplate implements CellTemplate<string> {
+export class TextCellTemplate implements CellTemplate<string> {
     readonly hasEditMode = true;
 
     validate(data: any): string {
-        const time_regex = /^\d{2}\:\d{2}$/;
-        data = data ? data.toString().replace(/\s+/g, '') : data;
-        return (time_regex.test(data)) ? data : '';
+        return (typeof (data) === 'string') ? data : '';
     }
 
     textToCellData(text: string): string {
@@ -23,31 +21,31 @@ export class TimeCellTemplate implements CellTemplate<string> {
         return { editable: true, cellData }
     }
 
-    customStyle: React.CSSProperties = {};
+    customStyle: React.CSSProperties = { background: '#fff' };
 
     renderContent: (props: CellRenderProps<string>) => React.ReactNode = (props) => {
         if (!props.isInEditMode)
             return props.cellData;
         const preserveValueKeyCodes = [0, keyCodes.ENTER];
         return <input
-            type='time'
             style={{
+                position: 'inherit',
                 width: '100%',
                 height: '100%',
                 padding: 0,
                 border: 0,
                 background: 'transparent',
                 fontSize: 14,
-                outline: 'none'
+                outline: 'none',
             }}
             ref={input => {
                 if (input) {
                     input.focus();
-                    // input.setSelectionRange(input.value.length, input.value.length);
+                    input.setSelectionRange(input.value.length, input.value.length);
                 }
             }}
             defaultValue={preserveValueKeyCodes.includes(props.lastKeyCode) ? props.cellData : ''}
-            onChange={e => props.onCellDataChanged ? props.onCellDataChanged(e.currentTarget.value) : null}
+            onChange={e => props.onCellDataChanged(e.currentTarget.value)}
             onCopy={e => e.stopPropagation()}
             onCut={e => e.stopPropagation()}
             onPaste={e => e.stopPropagation()}
