+++ src/lib/index.ts
@@ -1,0 +1,5 @@
+// This modules are exported by the lib
+export * from './Components/ReactGrid'
+export * from './Common/PublicModel'
+// TODO Range is INTERNAL! should not be public
+export * from './Common/Range'
