+++ tsconfig.json
@@ -24,5 +24,8 @@
   },
   "include": [
     "src"
+  ],
+  "exclude": [
+    "src/index.tsx"
   ]
 }
